chore: update documentation, fix unit tests, move files
